Refactored out methods that's not specific for ems
We will use the same methods in a later commit for
fetching feedbacks. Also some of the methods was
use a lot of other places in the codebase that didn't
have any direct connection to the ems class.
